# clean-architecture/docs/swagger/swagger.yaml
basePath: /api
definitions:
  github_com_Tomelin_fc-desafio-db_clean-architecture_internal_core_entity.Category:
    enum:
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-varnames:
    - Software
    - HomeAutomation
    - BigData
    - Unknow
  github_com_Tomelin_fc-desafio-db_clean-architecture_internal_core_entity.Order:
    properties:
      amount:
        type: integer
      category:
        $ref: '#/definitions/github_com_Tomelin_fc-desafio-db_clean-architecture_internal_core_entity.Category'
      description:
        type: string
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
    required:
    - amount
    - category
    - name
    - price
    - stock
    type: object
info:
  contact:
    email: contato@FullCycle.com.br
    name: FullCycle
    url: www.FullCycle.com.br
  description: desafio fullcycle
  title: manager orders
  version: 1.0.0
paths:
  /order:
    get:
      consumes:
      - application/json
      description: get all orders
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_Tomelin_fc-desafio-db_clean-architecture_internal_core_entity.Order'
            type: array
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: find all orders
      tags:
      - order
  /order/{id}:
    get:
      consumes:
      - application/json
      description: get order by ID
      parameters:
      - description: found
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Tomelin_fc-desafio-db_clean-architecture_internal_core_entity.Order'
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: find order by ID
      tags:
      - order
  /order/search:
    get:
      consumes:
      - application/json
      description: get order by filter
      parameters:
      - description: filter field
        in: query
        name: filter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_Tomelin_fc-desafio-db_clean-architecture_internal_core_entity.Order'
            type: array
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: find order by filter
      tags:
      - order
schemes:
- http
- https
swagger: "2.0"